requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
google-genai>=0.3.0
gunicorn==21.2.0
whitenoise==6.6.0
